+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Possible log types:
 
 ### Unreleased
 
+- [added] Add an `InMemoryPublicKeyCache` backed by the moka crate. Enabled through the
+  `public-key-cache-inmemory` Cargo feature (enabled by default).
+- [changed] The `cache` module is now exposed directly. The `PublicKeyCache` trait will not be re-exported at
+  top level anymore in the next breaking release.
+
 ### v0.19.0 (2026-01-04)
 
 - [added] Add support for generic public key caching:`lookup_pubkey_with_cache`